Fix auxiliary annotation tags and smooth canvas zoom

+++ composeApp/src/commonMain/kotlin/com/xiehe/spine/ui/components/analysis/image/ImageViewport.kt
@@ -15,7 +15,7 @@ fun ImageViewport(
     activeToolId: String,
     pendingPoints: List<MeasurementPoint>,
     isImageLocked: Boolean,
-    zoomPercent: Int,
+    zoomPercent: Float,
     contrast: Int,
     brightness: Int,
     onCanvasTap: (MeasurementPoint) -> Unit,
