+++ composeApp/src/commonMain/kotlin/com/xiehe/spine/ui/components/analysis/viewer/ImageViewer.kt
@@ -14,7 +14,7 @@ fun ImageViewer(
     activeToolId: String,
     pendingPoints: List<MeasurementPoint>,
     isImageLocked: Boolean,
-    zoomPercent: Int,
+    zoomPercent: Float,
     contrast: Int,
     brightness: Int,
     onCanvasTap: (MeasurementPoint) -> Unit,
